refactor(site-url): move substitution out of vite plugin into a post-build script
The previous approach relied on a Vite closeBundle hook inside vocs.config.ts
to rewrite files like docs/public/SKILL.md after vocs build wrote them to
dist/. It worked on local pnpm build but did NOT fire in Vercel's native build
environment — the preview deployment still served /SKILL.md with raw
__SITE_URL__ sentinels. Rather than debug Vocs's multi-stage Vite pipeline
(resolveVocsConfig, multiple vite.build calls, prerender phase, root set to
node_modules, etc.), just stop relying on Vite plugin ordering:

- scripts/site-url.mjs: single source of truth for the siteUrl chain
  (SITE_URL → VERCEL_PROJECT_PRODUCTION_URL → VERCEL_BRANCH_URL → VERCEL_URL
  → localhost fallback), imported from both vocs.config.ts (dev mode) and
  the post-build script (prod build).

- scripts/substitute-site-url.mjs: walks docs/dist after vocs build exits
  and rewrites every .md/.txt/.json/.html/.js file that still contains the
  sentinel. Logs a count when done so it's easy to verify in CI output.

- package.json: build script is now `vocs build && node
  scripts/substitute-site-url.mjs`. Guaranteed ordering — the post-build
  step runs after vocs build has written everything, regardless of what
  Vite's internal plugin order does.

- vocs.config.ts: drops the closeBundle hook entirely. Keeps only the
  transform hook, which is a dev-mode convenience so `pnpm dev` renders
  http://localhost:5173 in the prompt blocks without a full rebuild. The
  prod build doesn't depend on it anymore — the post-build script is
  authoritative.

Verified locally across all three env scenarios (no env vars, preview via
VERCEL_BRANCH_URL, production via VERCEL_PROJECT_PRODUCTION_URL) — all three
show zero sentinel leaks and correct URLs in SKILL.md, llms-full.txt,
quickstart/agent.md, and HTML output.
