+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vocs dev",
-    "build": "vocs build",
+    "build": "vocs build && node scripts/substitute-site-url.mjs",
     "preview": "vocs preview"
   },
   "dependencies": {
